Re-factored code for saving/loading to use standard KSP persistence (ScenarioModule in persistence file).
Benefits:
- standard persistence, no external file
- flight reverting leads to kerbal deaths / vessels destroyed being properly not registered for current review
