+++ Instance/StateFundingScenario.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace StateFunding
+{
+  public class StateFundingScenario
+  {
+    public StateFundingScenario ()
+    {
+    }
+  }
+}
+
