+++ StateFunding.version
@@ -1,0 +1,31 @@
+{
+     "NAME":"StateFunding",
+     "URL":"https://github.com/TheDogKSP/StateFunding/raw/master/StateFunding.version",
+     "DOWNLOAD":"https://github.com/TheDogKSP/StateFunding/releases",
+     "GITHUB":{
+         "USERNAME":"TheDogKSP",
+         "REPOSITORY":"StateFunding",
+         "ALLOW_PRE_RELEASE":false
+     },
+     "VERSION":{
+         "MAJOR":0,
+         "MINOR":4,
+         "PATCH":0,
+         "BUILD":0
+     },
+     "KSP_VERSION":{
+         "MAJOR":1,
+         "MINOR":1,
+         "PATCH":0
+     },
+     "KSP_VERSION_MIN":{
+         "MAJOR":1,
+         "MINOR":1,
+         "PATCH":0
+     },
+     "KSP_VERSION_MAX":{
+         "MAJOR":1,
+         "MINOR":1,
+         "PATCH":0
+     }
+ }
